Move fs-dependent periodic tasks into dump-manager.
