+++ lsif/src/dump-manager/dump-manager.ts
@@ -9,6 +9,11 @@ import { Logger } from 'winston'
 import { startExpressApp } from '../shared/api/init'
 import { createDatabaseRouter } from './routes/database'
 import { createUploadRouter } from './routes/uploads'
+import { startTasks } from './tasks'
+import { createPostgresConnection } from '../shared/database/postgres'
+import { waitForConfiguration } from '../shared/config/config'
+import { DumpManager } from '../shared/store/dumps'
+import { UploadManager } from '../shared/store/uploads'
 
 /**
  * No-op dump-manager process.
@@ -18,6 +23,9 @@ import { createUploadRouter } from './routes/uploads'
 async function main(logger: Logger): Promise<void> {
     // Collect process metrics
     promClient.collectDefaultMetrics({ prefix: 'lsif_' })
+
+    // Read configuration from frontend
+    const fetchConfiguration = await waitForConfiguration(logger)
 
     // Update cache capacities on startup
     metrics.connectionCacheCapacityGauge.set(settings.CONNECTION_CACHE_CAPACITY)
@@ -29,6 +37,14 @@ async function main(logger: Logger): Promise<void> {
     await ensureDirectory(path.join(settings.STORAGE_ROOT, constants.DBS_DIR))
     await ensureDirectory(path.join(settings.STORAGE_ROOT, constants.TEMP_DIR))
     await ensureDirectory(path.join(settings.STORAGE_ROOT, constants.UPLOADS_DIR))
+
+    // Create database connection and entity wrapper classes
+    const connection = await createPostgresConnection(fetchConfiguration(), logger)
+    const dumpManager = new DumpManager(connection)
+    const uploadManager = new UploadManager(connection)
+
+    // Start background tasks
+    startTasks(connection, dumpManager, uploadManager, logger)
 
     const routers = [createDatabaseRouter(logger), createUploadRouter(logger)]
