+++ lsif/src/dump-manager/settings.ts
@@ -18,3 +18,18 @@ export const DOCUMENT_CACHE_CAPACITY = readEnvInt('DOCUMENT_CACHE_CAPACITY', 102
 
 /** The maximum number of result chunks that can be held in memory at once. */
 export const RESULT_CHUNK_CACHE_CAPACITY = readEnvInt('RESULT_CHUNK_CACHE_CAPACITY', 1024 * 1024 * 1024)
+
+/** The interval (in seconds) to clean the dbs directory. */
+export const PURGE_OLD_DUMPS_INTERVAL = readEnvInt('PURGE_OLD_DUMPS_INTERVAL', 60 * 30)
+
+/** How many uploads to query at once when determining if a db file is unreferenced. */
+export const DEAD_DUMP_CHUNK_SIZE = readEnvInt('DEAD_DUMP_CHUNK_SIZE', 100)
+
+/** The maximum space (in bytes) that the dbs directory can use. */
+export const DBS_DIR_MAXIMUM_SIZE_BYTES = readEnvInt('DBS_DIR_MAXIMUM_SIZE_BYTES', 1024 * 1024 * 1024 * 10)
+
+/** The interval (in seconds) to invoke the cleanFailedUploads task. */
+export const CLEAN_FAILED_UPLOADS_INTERVAL = readEnvInt('CLEAN_FAILED_UPLOADS_INTERVAL', 60 * 60 * 8)
+
+/** The maximum age (in seconds) that the files for an unprocessed upload can remain on disk. */
+export const FAILED_UPLOAD_MAX_AGE = readEnvInt('FAILED_UPLOAD_MAX_AGE', 24 * 60 * 60)
